🎨 Cleaning up Navbar (#1)
* 🎨 Refactor icon code

* 🎨 Make icons individual components

* 🎨 Make folder for navbar component

* 🎨 Create circuit sub menu items component

* 🎨 Create sub components for navbar

* 🎨 Create mobile nav sub component

* ⚡️ Use Next.js image component

* 🔒 Update yarn lock

* 🎨 Clean up main navbar component

* 🎨 Create circuit sub menu component

* 🔀 Merge from main
