+++ components/Icons/ArchiveIcon.jsx
@@ -1,0 +1,33 @@
+function ArchiveIcon({ active }) {
+  return (
+    <svg
+      className="mr-2 h-5 w-5"
+      aria-hidden="true"
+      viewBox="0 0 20 20"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <rect
+        x="5"
+        y="8"
+        width="10"
+        height="8"
+        fill={active ? "#8B5CF6" : "#EDE9FE"}
+        stroke={active ? "#C4B5FD" : "#A78BFA"}
+        strokeWidth="2"
+      />
+      <rect
+        x="4"
+        y="4"
+        width="12"
+        height="4"
+        fill={active ? "#8B5CF6" : "#EDE9FE"}
+        stroke={active ? "#C4B5FD" : "#A78BFA"}
+        strokeWidth="2"
+      />
+      <path d="M8 12H12" stroke="#A78BFA" strokeWidth="2" />
+    </svg>
+  );
+}
+
+export default ArchiveIcon;
