+++ components/Icons/EditIcon.jsx
@@ -1,0 +1,20 @@
+function EditIcon({ active }) {
+  return (
+    <svg
+      className="mr-2 h-5 w-5 text-violet-400"
+      aria-hidden="true"
+      viewBox="0 0 20 20"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M4 13V16H7L16 7L13 4L4 13Z"
+        fill={active ? "#8B5CF6" : "#EDE9FE"}
+        stroke={active ? "#C4B5FD" : "#A78BFA"}
+        strokeWidth="2"
+      />
+    </svg>
+  );
+}
+
+export default EditIcon;
