+++ components/Icons/index.js
@@ -1,0 +1,7 @@
+import ArchiveIcon from "./ArchiveIcon";
+import DeleteIcon from "./DeleteIcon";
+import EditIcon from "./EditIcon";
+import DuplicateIcon from "./DuplicateIcon";
+import MoveIcon from "./MoveIcon";
+
+export { ArchiveIcon, DeleteIcon, EditIcon, DuplicateIcon, MoveIcon };
