+++ components/Navbar/MobileNav/index.js
@@ -1,0 +1,3 @@
+import MobileNav from "./MobileNav";
+
+export default MobileNav;
